requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-filter==25.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
pandas==2.2.2
openpyxl==3.1.2
numpy==1.26.4
python-dateutil==2.8.2
